Move tests alongside source and add CI
