+++ netlify/functions/schedule-notification-background.test.ts
@@ -4,13 +4,13 @@ const sub = { endpoint: 'x' }
 
 describe('schedule-notification-background', () => {
   it('returns 400 if no body', async () => {
-    const { handler } = await import('../netlify/functions/schedule-notification-background')
+    const { handler } = await import('./schedule-notification-background')
     const res = await handler({} as any)
     expect(res.statusCode).toBe(400)
   })
 
   it('returns 500 if missing keys', async () => {
-    const { handler } = await import('../netlify/functions/schedule-notification-background')
+    const { handler } = await import('./schedule-notification-background')
     const res = await handler({ body: JSON.stringify({ subscription: sub, payload: {}, delay: 0 }) } as any)
     expect(res.statusCode).toBe(500)
   })
@@ -21,7 +21,7 @@ describe('schedule-notification-background', () => {
     const webpush = await import('web-push')
     const sendNotification = vi.spyOn(webpush.default, 'sendNotification').mockResolvedValue(undefined as any)
     vi.spyOn(webpush.default, 'setVapidDetails').mockImplementation(() => {})
-    const { handler } = await import('../netlify/functions/schedule-notification-background')
+    const { handler } = await import('./schedule-notification-background')
     const res = await handler({ body: JSON.stringify({ subscription: sub, payload: { hi: 1 }, delay: 0 }) } as any)
     expect(sendNotification).toHaveBeenCalled()
     expect(res.statusCode).toBe(200)
