+++ src/App.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
 import { vi } from 'vitest'
-vi.mock('../src/assets/bus.png', () => ({ default: '' }))
-import App from '../src/App'
+vi.mock('./assets/bus.png', () => ({ default: '' }))
+import App from './App'
 
 describe('App', () => {
   it('renders without crashing', () => {
